refactor: SelectAbilityBox 컴포넌트명을 구별을 위해 NewSelectAbilityBox로 변경
기존에 역량 페이지에 있던 SelectAbilityBox 컴포넌트를 기반으로 NewSelectAbilityBox 컴포넌트를 별도로 만듬.
